fix compilation warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@18353 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a87994773b64ef02e5eaed6aebbdfa8402e1d799

+++ Rec/Tf/TrackSeeding/src/TrackSeedFind.cpp
@@ -2,9 +2,7 @@
 
 // from Gaudi
 #include "GaudiKernel/AlgFactory.h"
-#include "GaudiKernel/SmartIF.h"
-#include "GaudiKernel/IDetDataSvc.h"
-#include "GaudiKernel/IRegistry.h"
+
 // Math definitions
 #include "GaudiKernel/Transform3DTypes.h"
 #include "GaudiKernel/Point3DTypes.h"
@@ -391,7 +389,7 @@ StatusCode TrackSeedFind::getData() {
       if ( 0 == OTTimes) {
           warning() << " No OTTimes container! " << endreq;
           warning() << " continuing without OT " << endreq;
-      } else if (OTTimes->size()>m_maxOTHits) {
+      } else if ((int)OTTimes->size()>m_maxOTHits) {
          warning() << " Too many OT hits, too high multiplicity " << OTTimes->size() << endreq; 
          warning() << " Skipping OT for this event " << endreq;
 	 // return StatusCode::SUCCESS;
